Hw 19.1

+++ docker-compose-e2e.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -e
 source ./docker.properties
 export COMPOSE_PROFILES=test
 export PROFILE=docker
@@ -7,27 +8,46 @@ export PREFIX="${IMAGE_PREFIX}"
 export ALLURE_DOCKER_API=http://allure:5050/
 export HEAD_COMMIT_MESSAGE="local build"
 export ARCH=$(uname -m)
+export BROWSER=${1:-chrome}
+export BUILD_MODE=${2:-nobuild}
 
 docker compose down
 docker_containers=$(docker ps -a -q)
 docker_images=$(docker images --format '{{.Repository}}:{{.Tag}}' | grep 'niffler')
+required_docker_images=$(docker compose config --images)
+need_build=false
 
-if [ ! -z "$docker_containers" ]; then
-  echo "### Stop containers: $docker_containers ###"
-  docker stop $docker_containers
-  docker rm $docker_containers
-fi
+echo "Browser: $BROWSER"
+echo "Build mode: $BUILD_MODE"
 
-if [ ! -z "$docker_images" ]; then
-  echo "### Remove images: $docker_images ###"
-  docker rmi $docker_images
+if [ "$BUILD_MODE" = "build" ]; then
+          need_build=true
 fi
 
 echo '### Java version ###'
 java --version
-bash ./gradlew clean
-bash ./gradlew jibDockerBuild -x :niffler-e-2-e-tests:test -Duser.timezone=UTC
+for image in $required_docker_images; do
+  if ! docker image inspect "$image" >/dev/null 2>&1; then
+        echo "Missing: $image"
+        need_build=true
+      fi
+done
 
-docker pull twilio/selenoid:chrome_stable_140
+if [ "$need_build" = true ]; then
+  if [ ! -z "$docker_images" ]; then
+    echo "### Remove images: $docker_images ###"
+    docker rmi $docker_images
+  fi
+  echo "### Build images ###"
+  bash ./gradlew clean
+  bash ./gradlew jibDockerBuild -x :niffler-e-2-e-tests:test -Duser.timezone=UTC
+else
+  echo "### Images already exist, skip build ###"
+fi
+if [ "$BROWSER" = "firefox" ]; then
+  docker image inspect selenoid/vnc_firefox:125.0 >/dev/null 2>&1 || docker pull selenoid/vnc_firefox:125.0
+else
+  docker image inspect twilio/selenoid:chrome_stable_140 >/dev/null 2>&1 || docker pull twilio/selenoid:chrome_stable_140
+fi
 docker compose up -d
 docker ps -a
